+++ models/contacts.json
@@ -58,11 +58,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "A",
-    "name": "Alen ",
-    "email": "nulla",
-    "phone": "914-3792"
   }
 ]
